requirements.txt
pytest==4.2.0
PyYAML==3.13
requests==2.21.0